pyproject.toml
[project]
name = "al-tools"
version = "0.1.0"
description = "Tooling for AnkiLangs.org"
authors = [
    { name = "Omar Kohl", email = "omarkohl@posteo.net" }
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "brain-brew>=0.3.11",
    "pandas>=2.3.1",
    "google-cloud-texttospeech>=2.27.0",
    "pyyaml>=6.0",
    "xlsxwriter>=3.2.0",
]

[dependency-groups]
dev = [
    "ruff>=0.12.7",
    "pytest>=9.0.3",
    "openpyxl>=3.1.5",
    "pytest-cov>=7.0.0",
]

[project.scripts]
al-tools = "al_tools.cli:cli"

[tool.pytest.ini_options]
addopts = "--cov=al_tools --cov-report=term-missing --cov-fail-under=50"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
